[mkt] revie apply: rest api errors
