+++ core/market/decentralized/src/negotiation/errors.rs
@@ -1,10 +1,7 @@
-use actix_web::ResponseError;
 use thiserror::Error;
 
 #[derive(Error, Debug)]
 pub enum NegotiationError {}
-
-impl ResponseError for NegotiationError {}
 
 #[derive(Error, Debug)]
 pub enum NegotiationInitError {}
